Move route functions into controller files and add starability css/html to replace rating slider.
