+++ models/review.js
@@ -4,6 +4,10 @@ const { Schema } = mongoose;
 const reviewSchema = new Schema({
   body: String,
   rating: Number,
+  author: {
+    type: Schema.Types.ObjectId,
+    ref: 'User'
+  }
 }) 
 
 module.exports = mongoose.model('Review', reviewSchema)
