+++ package.json
@@ -18,7 +18,8 @@
     "joi": "^17.6.0",
     "method-override": "^3.0.0",
     "mongoose": "^6.3.6",
-    "passport": "^0.6.0",
+    "nodemon": "^2.0.18",
+    "passport": "^0.5.0",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^7.1.2"
   }
