+++ views/campgrounds/show.ejs
@@ -1,4 +1,6 @@
 <% layout('layouts/boilerplate') %> 
+<link rel="stylesheet" href="/stylesheets/stars.css">
+
 <div class="row">
   <div class="col-6">
     <div class="card mb-4">
@@ -9,43 +11,73 @@
       </div>
       <ul class="list-group list-group-flush">
         <li class="list-group-item text-muted"><%= campground.location %></li>
+        <li class="list-group-item">Submitted By <%= campground.author.username %></li>
         <li class="list-group-item">$<%= campground.price %> /Night</li>
       </ul>
-      <div class="card-body">
-        <a href="/campgrounds/<%= campground._id %>/edit" class="card-link btn btn-warning">Edit</a>
-        <form class="d-inline" action="/campgrounds/<%= campground._id %>?_method=DELETE" method="POST">
-          <button class="btn btn-danger">Delete</button>
-        </form>
-        <a href="/campgrounds" class="card-link btn btn-primary">All Campgrounds</a>
-      </div>
+
+      <% if(currentUser && campground.author.equals(currentUser)) {%> 
+        <div class="card-body">
+          <a href="/campgrounds/<%= campground._id %>/edit" class="card-link btn btn-warning">Edit</a>
+          <form class="d-inline" action="/campgrounds/<%= campground._id %>?_method=DELETE" method="POST">
+            <button class="btn btn-danger">Delete</button>
+          </form>
+          <a href="/campgrounds" class="card-link btn btn-primary">All Campgrounds</a>
+        </div>
+      <% } %> 
+
     </div>
   </div>
   <div class="col-6">
-    <h2>Leave a Review</h2>
-    <form class="mb-3 validated-form" action="/campgrounds/<%= campground._id%>/reviews" method="POST" novalidate>
-      <div class="mb-3">
-        <label class="form-label" for="rating">Rating</label>
-        <input class="form-range" type="range" min="1" max="5" name="review[rating]" id="rating">
-      </div>
-      <div class="mb-3">
-        <label class="form-label" for="body">Review</label>
-        <textarea class="form-control" name="review[body]" id="body" cols="30" rows="3" required></textarea>
-        <div class="valid-feedback">
-          Looks good!
+
+    <% if(currentUser) { %> 
+      <h2>Leave a Review</h2>
+      <form class="mb-3 validated-form" action="/campgrounds/<%= campground._id%>/reviews" method="POST" novalidate>
+
+        <div class="mb-1">
+          <fieldset class="starability-basic">
+            <input type="radio" id="no-rate" class="input-no-rate" name="review[rating]" value="1" checked aria-label="No rating." />
+            <input type="radio" id="first-rate1" name="review[rating]" value="1" />
+            <label for="first-rate1" title="Terrible">1 star</label>
+            <input type="radio" id="first-rate2" name="review[rating]" value="2" />
+            <label for="first-rate2" title="Not good">2 stars</label>
+            <input type="radio" id="first-rate3" name="review[rating]" value="3" />
+            <label for="first-rate3" title="Average">3 stars</label>
+            <input type="radio" id="first-rate4" name="review[rating]" value="4" />
+            <label for="first-rate4" title="Very good">4 stars</label>
+            <input type="radio" id="first-rate5" name="review[rating]" value="5" />
+            <label for="first-rate5" title="Amazing">5 stars</label>
+          </fieldset>
         </div>
-      </div>
-      <button class="btn btn-success">Submit</button>
-    </form>
+
+        <div class="mb-3">
+          <label class="form-label" for="body">Review</label>
+          <textarea class="form-control" name="review[body]" id="body" cols="30" rows="3" required></textarea>
+          <div class="valid-feedback">
+            Looks good!
+          </div>
+        </div>
+        <button class="btn btn-success">Submit</button>
+      </form>
+    <% } %> 
 
     <% for( let review of campground.reviews ) { %>
       <div class="card mb-3">
         <div class="card-body">
-          <h5 class="card-title">Rating: <%= review.rating %></h5>
+
+          <p class="starability-result" data-rating="<%= review.rating %>">
+            Rated: <%= review.rating %>  stars
+          </p>
+          
+          <h6 class="card-subtitle text-muted mb-2">By <%= review.author.username %></h6>
           <p class="card-text">Review: <%= review.body %></p>
-          <form action="/campgrounds/<%= campground._id %>/reviews/<%= review._id %>?_method=DELETE" 
-            method="POST">
-            <button class="btn btn-sm btn-danger">Delete</button>
-          </form>
+
+          <% if(currentUser && review.author.equals(currentUser)) {%> 
+            <form action="/campgrounds/<%= campground._id %>/reviews/<%= review._id %>?_method=DELETE" 
+              method="POST">
+              <button class="btn btn-sm btn-danger">Delete</button>
+            </form>
+          <% } %> 
+
         </div>
     </div>
     <% } %>
